requirements.txt
streamlit         # Core app framework
pandas            # Data handling
openpyxl          # Excel support (.xlsx)
altair            # Charts (used in HORPT2 etc.)
plotly            # Optional charts (e.g. px.bar)
numpy             # General numeric support
gspread>=5.7.0
google-auth>=2.15.0
pandas>=1.3.0
streamlit>=1.28.0